app/server/requirements.txt
django==6.0.4
djangorestframework==3.17.1
psycopg[binary]==3.3.3
gunicorn==23.0.0
whitenoise==6.12.0
